Refactor to decouple ButtonSequence
