+++ mdpicgen.py
@@ -1,4 +1,6 @@
-from extract_md import extract_button_sequences, format_image_basename, ButtonSequence
+from extract_md import extract_button_sequences
+from util import format_image_basename
+from button_sequence import ButtonSequence
 from modify_md import format_markdown, write_markdown
 from imageset_gen import ImageSet, ImageOpt
 from psd_gen import PSDInMd
